\name{Families-package}
\alias{Families-package}
\alias{FamiliesPop}
\docType{package}
\title{
\packageTitle{Families}
}
\description{
\packageDescription{Families}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{Families}
\packageIndices{Families}
}
\author{Frans Willekens <Willekens@nidi.nl>}
\keyword{ Family demography}
