% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDau.R
\name{IDau}
\alias{IDau}
\title{IDs of aunts and uncles of ego}
\usage{
IDau(idego, d = NULL)
}
\arguments{
\item{idego}{Identification number(s) of ego(s).}

\item{d}{Name of database. If missing, dLH is used, if it exists in the global environment (i.e. R workspace). 
Otherwise the dataset dLH distributed with the Families package are used.}
}
\value{
IDs of aunts and uncles by blood (without their partner, i.e. in-laws).
}
\description{
Retrieves the IDs of ego's aunts and uncles by blood. Partners (in-laws) are not included. 
They are obtained using the IDPartner() function.
}
\examples{
# Load the data
data(dLH,package="Families")
set.seed <- 45
idego <- sample(dLH$ID[dLH$gen==3],1)
idau <- IDau (idego)

idego <- dLH$ID[dLH$gen==3]
idau <- IDau (idego,d=dLH)

}
