% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nkin.R
\name{Nkin}
\alias{Nkin}
\title{Number of kin of given type}
\usage{
Nkin(idkin)
}
\arguments{
\item{idkin}{dataframe of IDs of ego and IDs of kin of given type}
}
\value{
Number of kin of given type
}
\description{
Computes, for each ego, number of kin of given type (parents, grandparents, aunts, uncles, siblings, etc.)
}
\examples{
# Load the data
data(dLH,package="Families")
idego <- dLH$ID[dLH$gen==3]
## Number of siblings
idsib <- IDsib(idego)
names(idsib$id) <- idego
nsib <- Nkin(idkin=idsib$d)

}
