% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Db.R
\name{Db}
\alias{Db}
\title{Date(s) of birth in decimal format}
\usage{
Db(idego, d = NULL)
}
\arguments{
\item{idego}{vector of IDs of egos}

\item{d}{Name of database. If d is missing, dLH is used.}
}
\value{
Dates of birth (decimal format)
}
\description{
Retrieves date(s) of birth from the database
}
\examples{
# Load the data
data(dLH,package = "Families")

# Date of birth of first individual in database
d <- dLH
Db(idego=1,d=d) 


}
