% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Db.R
\name{Db}
\alias{Db}
\title{Retrieves the date(s) of birth in decimal format}
\usage{
Db(idego, dLH)
}
\arguments{
\item{idego}{vector of IDs of egos}

\item{dLH}{Name of database. If dLH is missing, dataLH_F is used.}
}
\value{
Returns the dates of birth
}
\description{
Retrieves the date(s) of birth from the database
}
\examples{

# Date of birth of first individual in database
data(dataLH_F,package = "Families")
Db(idego=1) 

}
\author{
Frans Willekens
}
