% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dd.R
\name{Dd}
\alias{Dd}
\title{Date(s) of death in decimal format}
\usage{
Dd(idego, d = NULL)
}
\arguments{
\item{idego}{vector of IDs of egos}

\item{d}{Name of database. If d is missing, dLH is used.}
}
\value{
Date of death (decimal format)
}
\description{
Retrieves date(s) of death from the database
}
\examples{
# Load the data
data(dLH,package = "Families")

# Date of death of first individual in database
Dd(idego=1,d=dLH) 

}
