% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dd.R
\name{Dd}
\alias{Dd}
\title{Retrieves the date(s) of death in decimal format}
\usage{
Dd(idego, dLH)
}
\arguments{
\item{idego}{vector of IDs of egos}

\item{dLH}{Name of database. If dLH is missing, dataLH_F is used.}
}
\value{
Returns the date of death
}
\description{
Retrieves the date(s) of death from the database
}
\examples{

# Date of death of first individual in database
data(dataLH_F,package = "Families")
Dd(idego=1) 

}
\author{
Frans Willekens
}
