% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e0.R
\name{e0}
\alias{e0}
\title{Computes (a) Life expectancy at birth, (b) Probability of surviving at age
65, and (c) Probability of surviving at age 85}
\usage{
e0(dLH)
}
\arguments{
\item{dLH}{The name of the database. If missing, dataLH_F is used.}
}
\value{
\item{e0}{Mean ages at death} \item{Prob65}{Probability of surviving
at age 65} \item{Prob85}{Probability of surviving at age 85}
}
\description{
Computes (a) Life expectancy at birth, (b) Probability of surviving at age
65, and (c) Probability of surviving at age 85
}
\examples{

data(dataLH_F,package = "Families")
e0(dLH=dataLH_F)

}
\author{
Frans Willekens
}
