% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Families-package.R
\docType{package}
\name{Families-package}
\alias{Families-package}
\alias{FamiliesPop}
\title{Kinship Ties in Virtual Populations}
\description{
Tools to study kinship networks, grandparenthood, and double burden (presence of children and oldest old parents) in virtual population produced by ’VirtualPop’.
}
\author{
Frans Willekens <Willekens@nidi.nl>
}
\keyword{demography}
\keyword{family}
\keyword{kinship}
