% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDfather.R
\name{IDfather}
\alias{IDfather}
\title{ID of father of ego}
\usage{
IDfather(idego, d = NULL)
}
\arguments{
\item{idego}{ID of ego(s)}

\item{d}{Name of database. If missing the dataset dLH distributed with the Families package is used.}
}
\value{
ID of father or (if keep_ego=TRUE, dataframe with ego-father dyals: ID of ego and ID of
father). Returns NA if ID of father is not included in the database
}
\description{
Retrieves the ID of the father of ego or fathers of vector of egos. ID of the father is the
ID of the partner of the mother. If the ID of the father listed in the dataset d differs from 
the ID of the mother's partner, a warning is given.
}
\examples{
# Load the data
data(dLH,package = "Families")

set.seed(31)
idf <- IDfather (idego=sample (dLH$ID[dLH$gen>=2],10))


}
