% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDfather.R
\name{IDfather}
\alias{IDfather}
\title{Retrieves ID of father of ego}
\usage{
IDfather(idego, dLH, keep_ego = FALSE)
}
\arguments{
\item{idego}{ID}

\item{dLH}{Name of database. If missing, dataLH_F is used.}

\item{keep_ego}{Option to link show ID of ego together with ID of father}
}
\value{
ID of father or (if keep_ego=TRUE, object with ID of ego and ID of
father). Returns NA if ID of father is not included in the database
}
\description{
Function to retrieve the ID of father of ego or fathers of vector of egos
}
\examples{

data(dataLH_F,package = "Families")
IDfather (idego=sample (dataLH_F$ID,10))

}
\author{
Frans Willekens
}
