% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDpartner.R
\name{IDpartner}
\alias{IDpartner}
\title{ID of partner of ego}
\usage{
IDpartner(idego, d = NULL)
}
\arguments{
\item{idego}{vector of IDs of egos.}

\item{d}{Name of database. If d is missing, the dataset dLH in the global environment (R workspace) is used. 
If no dLH in the global environment, the database dLH distributed with the Families package is used.}
}
\value{
Vector of IDs of partners
}
\description{
Retrieves ID of partners of vector of egos #'
}
\examples{
# Load the data
data(dLH,package = "Families")
# Get ID of partner
IDpartner(idego=1,d=dLH)
IDpartner(idego=c(4,9,NA,30),d=dLH)  

}
