% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests.R
\name{Tests}
\alias{Tests}
\title{Tests existence of data file.}
\usage{
Tests(idego, d = NULL)
}
\arguments{
\item{idego}{ID}

\item{d}{Name of database. If d is missing, the dataset dLH in the global environment (R workspace) is used. 
If no dLH in the global environment, the database dLH distributed with the Families package is used.}
}
\value{
idego and d
}
\description{
Tests existence of data file. If the data file d is null, the function checks whether dLH exists in the global environment. 
If it does, d=dLH. If not, the message "Object dLH is not provided and does not exist" is displayed and
the processing stops. 
If some values in idego vector are missing (are NA) a message is displayed and the missing elements are removed.
}
